requirements.txt
Django==1.11
Pillow==4.3.0
djangorestframework==3.7.1
boto3==1.4.7
django-storages==1.6.5
python-decouple==3.1
datetime==4.2
pytest-django==3.1.2
psycopg2==2.7.3.2
dj-database-url==0.4.2
firebase-admin==2.5.0
google-cloud-firestore==0.28.0
google-gax==0.15.16

